let user set the side_panel height
This feature is useful for apps that need to be run both on mobile
devices and on PCs. On the PC, the whole screen height might not be
needed, which leaves dead empty space in the current version.
With this new feature, the default height is the parent’s height, and
it can be explicitly set to be different by setting the
side_panel_height property to a custom value.